Adds persistent cacheing of SegYReader objects.
SegYReader objects are now cached by default in a persistent
store of pickles. Cached readers are located on the basis of
content hashes of the SEG Y files to which they relate combined
with the other arguments used to create the orginal reader. This
can lead to significant speed-ups since all indexes owned by the
reader instance are also cached.